refactor: Move coil packages out of image dir
