+++ app/src/main/java/eu/kanade/tachiyomi/data/coil/CoilDiskCache.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.data.image.coil
+package eu.kanade.tachiyomi.data.coil
 
 import android.content.Context
 import coil3.disk.DiskCache
