+++ app/src/main/java/eu/kanade/tachiyomi/data/coil/CoverViewTarget.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.data.image.coil
+package eu.kanade.tachiyomi.data.coil
 
 import android.graphics.drawable.Drawable
 import android.view.View
