+++ app/src/main/java/eu/kanade/tachiyomi/data/coil/LibraryMangaImageTarget.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.data.image.coil
+package eu.kanade.tachiyomi.data.coil
 
 import android.graphics.BitmapFactory
 import android.widget.ImageView
