+++ app/src/main/java/eu/kanade/tachiyomi/data/coil/Utils.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.data.image.coil
+package eu.kanade.tachiyomi.data.coil
 
 import coil3.Extras
 import coil3.getExtra
