+++ app/src/main/java/eu/kanade/tachiyomi/ui/extension/ExtensionHolder.kt
@@ -12,7 +12,7 @@ import androidx.core.view.isVisible
 import coil3.dispose
 import coil3.load
 import eu.kanade.tachiyomi.R
-import eu.kanade.tachiyomi.data.image.coil.CoverViewTarget
+import eu.kanade.tachiyomi.data.coil.CoverViewTarget
 import eu.kanade.tachiyomi.databinding.ExtensionCardItemBinding
 import eu.kanade.tachiyomi.extension.model.Extension
 import eu.kanade.tachiyomi.extension.model.InstallStep
